fix(proxy): redact Live sideband account selection
