+++ tests/unit/test_proxy_utils.py
@@ -5174,6 +5174,7 @@ async def test_select_codex_control_account_without_budget_uses_balancer(monkeyp
         secondary_budget_threshold_pct=100.0,
         traffic_class=proxy_service.TRAFFIC_CLASS_FOREGROUND,
         concurrency_caps=ANY,
+        redact_sensitive_details=False,
     )
